refactor: Move store registering screen
Move store registering screen
